requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
attrs==25.4.0
boltons==25.0.0
boto3==1.38.26
botocore==1.38.26
cachetools==5.5.2
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
Deprecated==1.2.18
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
furl==2.1.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-tools==1.71.0
gspread==5.12.2
h11==0.14.0
hf-xet==1.1.10
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.35.3
idna==2.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
legacy-cgi==2.6.3
magic-filter==1.0.12
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
orderedmultidict==1.0.1
packaging==25.0
pillow==12.0.0
propcache==0.4.1
protobuf==5.29.5
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyTelegramBotAPI==4.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==20.3
pytz==2025.2
PyYAML==6.0.2
regex==2025.10.22
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
Telethon==1.40.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.4.0
wrapt==1.17.2
yandexcloud==0.347.0
yappa==0.4.30
yarl==1.22.0
